fix(desktop): drop redundant f32 conversion in sidebar yield sync
clamp_sidebar now returns f32 directly (metrics batch 08e2fd4), so the
f32::from wrapper tripped clippy useless_conversion and blocked the
verify gate at HEAD. Remove the wrapper only; no behavior change, the
clippy gate is the regression lock.

+++ crates/dsh-desktop/src/shell/store.rs
@@ -974,7 +974,7 @@ impl AppStore {
         } else {
             0.
         };
-        let expanded_w = f32::from(crate::shell::metrics::clamp_sidebar(self.sidebar_px));
+        let expanded_w = crate::shell::metrics::clamp_sidebar(self.sidebar_px);
         let sidebar_unfit =
             viewport_w - expanded_w < crate::shell::metrics::CHAT_AREA_MIN + panel_reserve;
         if sidebar_unfit && !self.sidebar_collapsed {
